SEO titles: prefer "Old Norse" over "Cleasby & Vigfusson"
The sister project of this site "Old Icelandic Dictionary" seems to fare much better SEO-wise. It seems that generic "Old Icelandic Dictionary" is more approachable in search results than "Cleasby & Vigfusson Dictionary", even if "Old Norse" is often mentioned. In general, readers of ON seem to appreciate Cleasby & Vigfusson book over the Zoega one, so this should be more a question of people not finding it in search engines.

Keep main mentions of "Cleasby & Vigfusson" to be transparent about the data source, but use "Old Norse Dictionary" in SEO titles & matching lines in content. Should make it more generic in terms of search result, which may help in generic "x in old norse" queries

Fixes #239